Split unit tests from aggregation job driver file
